+++ hackage-ui/components/pages/package/Sidebar.tsx
@@ -13,6 +13,7 @@ import updatedAtIcon from '!!raw-loader!../../icons/updated-at.svg';
 import SidebarButton from "../../forms/SidebarButton";
 import { ExtA } from "../../layout/A";
 import { PackageProps } from './common';
+import { Repository, parseRepositoryUrl } from '../../api/Github';
 
 const tooltipId = 'package-sidebar-tooltip';
 
@@ -150,45 +151,6 @@ export const Sidebar = (props: SidebarProps) => {
       )}
     </div >
   );
-}
-
-type Repository = {
-  kind: 'unknown' | 'github',
-  displayText: string,
-  browserUrl: string,
-  gitUrl: string | null
-}
-
-// TODO - add tests, GitLab, BitBucket, others.
-function parseRepositoryUrl(url: string): Repository {
-
-  // GitHub
-  function gitHub(owner: string, repo: string): Repository {
-    return {
-      kind: 'github',
-      displayText: `${owner}/${repo}`,
-      browserUrl: `https://github.com/${owner}/${repo}`,
-      gitUrl: `git@github.com:${owner}/${repo}.git`
-    }
-  }
-
-  if (url.match(/^git@github.com:.*$/g)) {
-    const [owner, repo] = url.replace(/^git@github\.com:\/?/, '').replace(/\/$/, '').replace(/\.git$/, '').split('/');
-    return gitHub(owner, repo);
-
-  } else if (url.match(/^(https?|git)\:\/\/github.com\/.*$/g)) {
-    const [owner, repo] = url.replace(/^(https?|git)\:\/\/github\.com\/?/, '').replace(/\/$/, '').replace(/\.git$/, '').split('/');
-    return gitHub(owner, repo);
-
-  }
-
-  // Unknown
-  return {
-    kind: 'unknown',
-    displayText: url,
-    browserUrl: url,
-    gitUrl: url.match(/git\:\/\//) ? url : null
-  }
 }
 
 export default Sidebar;
